isDesktop false if isWeb true in responsive.dart
https://gitlab.com/veilid/veilidchat/-/tree/268b86d1319d3bd77dff857c7154250679a081ef
tries to access Platform._operatingSystem as part of checking whether
it is running on Windows, Linux, or macOS, and this information is not
available on the web. This MR makes it possible to reach the
VeilidChat "Create a new account" screen when running in Chrome
(although the user apparently can't proceed further because of an
"already borrowed" panic). Without this MR, there is an uncaught error
even before getting to that account screen.